# web/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
        "/notifications/send": {
            "post": {
                "summary": "Send a notification",
                "operationId": "088e9c35c886513afc6d109728d82295",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Hello, this is a test notification"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "email",
                                            "sms"
                                        ],
                                        "example": "sms"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Failure"
                    }
                }
            }
        },
        "/user/preferences": {
            "get": {
                "summary": "Get user preferences",
                "operationId": "4d33d3328eae150fa4bf17acaaa202e3",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "summary": "Update user preferences",
                "operationId": "5c03a7eccbc8711d5bb43d21110e1108",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notification_preference": {
                                        "type": "string",
                                        "enum": [
                                            "email",
                                            "sms"
                                        ],
                                        "example": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        }
    }
}